refactor : utils 파일 간소화
- 기존의 utils 폴더안의 파일들이 함수 단위로 구성되어 보기 복잡한 점을 고려하여 date 파일에 관련 함수들을 묶음
