+++ client/src/components/organisms/FourMonthStatistics/FourMonthStatistics.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import { getPastMonthList } from '@utils/calcMonth';
+import { getPastMonthList } from '@utils/date';
 import { numberToMoney } from '@utils/number';
 import getRandomKey from '@utils/random';
 import RowFlexContainer from '@atoms/div/RowFlexContainer';
